Add service type.
This allows deployments to re-use the same certificate across replicas.
The problem is that the node type uses the name of the node as the
certificate -- so "node-0", "node-1" etc. Which is maybe OK for static
pod names from a stateful set. However, for deployment replicas,
which have unique names like "myservice-5675675756", this would require
manual csr approvel every time a pod was recycled.

This tweak adds a `service` type and `-service-name=...` argument that
generates certificates that can be shared by multiple pods.